requirements.txt
asgiref==3.8.1
astroid==3.2.4
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.7.1
dill==0.3.8
dj-rest-auth==6.0.0
Django==5.0.7
django-allauth==0.61.1
django-cors-headers==4.4.0
django-debug-toolbar==4.4.5
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.7.1
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
isort==5.13.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pylint==3.2.6
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.18.1
sqlparse==0.5.0
tomlkit==0.13.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
